requirements.txt
Flask
Flask-RestPlus
Redis
rq
rq-scheduler
gunicorn
python-dotenv
requests
urlparse